docs :: to common dir
editing modal component and move to common dir
